fix: rename files and update references
