+++ integrations/jenkins/webhook_processors/__init__.py
@@ -1,4 +1,4 @@
-from webhook_processors.build_processor import BuildWebhookProcessor
-from webhook_processors.job_processor import JobWebhookProcessor
+from webhook_processors.build_webhook_processor import BuildWebhookProcessor
+from webhook_processors.job_webhook_processor import JobWebhookProcessor
 
 __all__ = ["BuildWebhookProcessor", "JobWebhookProcessor"]
